awallet/requirements.txt
absl-py==0.9.0
backcall==0.1.0
decorator==4.4.1
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycryptodome==3.9.4
Pygments==2.5.2
six==1.14.0
traitlets==4.3.3
wcwidth==0.1.8
yapf==0.29.0
